Make ctrl+s save
However, it does appear we cannot inject
Applicaions. So we will have to look at another
way of doing this while still maintaining somewhat
of a structure.
