+++ src/main/kotlin/view/main/EntriesView.kt
@@ -2,6 +2,7 @@ package main.kotlin.view.main
 
 import javafx.scene.control.ButtonType
 import javafx.scene.layout.Priority
+import main.kotlin.JournalApp
 import main.kotlin.Styles
 import main.kotlin.controller.EditorController
 import main.kotlin.controller.JournalController
@@ -11,17 +12,17 @@ import java.time.LocalDate
 import java.time.temporal.ChronoUnit
 
 class EntriesView : View() {
+    private val journalApp: JournalApp by inject()
+
     val editorController: EditorController by inject()
     val journalController: JournalController by inject()
-
-    val menuViewView: MenuView by inject()
 
     override val root = vbox {
         addClass(Styles.customContainer)
 
         text("Entries") { addClass(Styles.title) }
 
-        listview(journalController.journalProperty.select { it.entriesProperty }) {
+        listview(journalController.journalProperty.select { it.itemsProperty }) {
             cellFragment(EntryFragment::class)
             vgrow = Priority.ALWAYS
             hgrow = Priority.NEVER
@@ -49,15 +50,15 @@ class EntriesView : View() {
             }
             button("save") {
                 disableWhen(journalController.journalProperty.select { it.editedProperty.not() })
-                action { menuViewView.save() }
+                action { journalApp.save() }
             }
             button("+").action {
                 var selection = ButtonType.YES
 
-                if (journalController.journalProperty.isNotNull.get() && journalController.journalProperty.get().entriesProperty.isNotEmpty()) {
+                if (journalController.journalProperty.isNotNull.get() && journalController.journalProperty.get().itemsProperty.isNotEmpty()) {
                     val daysAfterEpoch = ChronoUnit.DAYS.between(LocalDate.ofEpochDay(0), LocalDate.now())
 
-                    val lastEntry = journalController.journalProperty.get().entriesProperty.last()
+                    val lastEntry = journalController.journalProperty.get().itemsProperty.last()
                     val lastDate = lastEntry.creationProperty.get()
                     val journalDaysAfterEpoch = ChronoUnit.DAYS.between(LocalDate.ofEpochDay(0), lastDate)
